Gaussian Quadrature

Gauss-Legendre quadrature

FastGaussQuadrature.gausslegendre — Method

gausslegendre(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Legendre quadrature.

\[\int_{-1}^{1} f(x) dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausslegendre(3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 2/5
true

Gauss-Hermite quadrature

FastGaussQuadrature.gausshermite — Method

gausshermite(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Hermite quadrature.

\[\int_{-\infty}^{+\infty} f(x) \exp(-x^2) dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausshermite(3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 3(√π)/4
true

Gauss-Laguerre quadrature

FastGaussQuadrature.gausslaguerre — Method

gausslaguerre(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Laguerre quadrature.

\[\int_{0}^{+\infty} f(x) e^{-x} dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausslaguerre(3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 24
true

FastGaussQuadrature.gausslaguerre — Method

gausslaguerre(n::Integer, α::Real) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of generalized Gauss-Laguerre quadrature.

\[\int_{0}^{+\infty} f(x) x^\alpha e^{-x} dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausslaguerre(3, 1.0);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 120
true

Optionally, a reduced quadrature rule can be computed. In that case, only those points and weights are computed for which the weight does not underflow in the floating point precision type. Supply the optional argument reduced = true.

Though the code is generic, heuristical choices on the choice of the algorithm are based on achieving machine precision accuracy only for Float64 type. In case the default choice of algorithm does not achieve the desired accuracy, the user can manually invoke the following routines:

gausslaguerre_GW: computation based on Golub-Welsch

gausslaguerre_rec: computation based on Newton iterations applied to evaluation using the recurrence relation

gausslaguerre_asy: the asymptotic expansions

Gauss-Chebyshev quadrature

FastGaussQuadrature.gausschebyshev — Method

gausschebyshev(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}
gausschebyshev(n::Integer, 1) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Chebyshev quadrature of the 1st kind.

\[\int_{-1}^{1} \frac{f(x)}{\sqrt{1-x^2}} dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausschebyshev(3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 3π/8
true

gausschebyshev(n::Integer, 2) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Chebyshev quadrature of the 2nd kind.

\[\int_{-1}^{1} f(x)\sqrt{1-x^2} dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausschebyshev(3, 2);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ π/16
true

gausschebyshev(n::Integer, 3) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Chebyshev quadrature of the 3rd kind.

\[\int_{-1}^{1} f(x)\sqrt{1-x^2} dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausschebyshev(3, 3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 3π/8
true

gausschebyshev(n::Integer, 4) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Chebyshev quadrature of the 4th kind.

\[\int_{-1}^{1} f(x)\sqrt{1-x^2} dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausschebyshev(3, 4);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 3π/8
true

Gauss-Jacobi quadrature

FastGaussQuadrature.gaussjacobi — Method

gaussjacobi(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Jacobi quadrature.

\[\int_{-1}^{1} f(x) (1-x)^\alpha(1+x)^\beta dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gaussjacobi(3, 1/3, -1/3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 268π/729(√3)
true

Gauss-Radau quadrature

FastGaussQuadrature.gaussradau — Method

gaussradau(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Radau quadrature.

\[\int_{-1}^{1} f(x) dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gaussradau(3);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 2/5
true

Note that the first node is fixed at -1.

julia> x, w = gaussradau(3);

julia> x[1]
-1.0

Gauss-Lobatto quadrature

FastGaussQuadrature.gausslobatto — Method

gausslobatto(n::Integer) -> Tuple{Vector{Float64},Vector{Float64}}

Return nodes and weights of Gauss-Lobatto quadrature.

\[\int_{-1}^{1} f(x) dx \approx \sum_{i=1}^{n} w_i f(x_i)\]

Examples

julia> x, w = gausslobatto(4);

julia> f(x) = x^4;

julia> I = dot(w, f.(x));

julia> I ≈ 2/5
true

Note that the both ends of nodes are fixed at -1 and 1.

julia> x, w = gausslobatto(4);

julia> x[1], x[end]
(-1.0, 1.0)