Misc.

About the logo

The center points of circles are generated from nodes(x-axis) and weights(y-axis) of Gaussian quadrature.

Number of points : n = 15

$\textcolor{#CB3C33}{\text{\textbf{Red}}}$ : gausslegendre(n)

$\textcolor{#389826}{\text{\textbf{Green}}}$ : gausschebyshev(n, 3)

$\textcolor{#9558B2}{\text{\textbf{Purple}}}$ : gaussjacobi(n, 5/2, 1/2)

These colors are from julia-logo-graphics.

The logo is generated with Luxor.jl, and its source code is here.

Other docstrings from private methods

FastGaussQuadrature.AIRY_ROOTS — Constant

The first 11 roots of the Airy function in Float64 precision https://mathworld.wolfram.com/AiryFunctionZeros.html

Examples

julia> zeros = airy.(FastGaussQuadrature.AIRY_ROOTS);

julia> all(zeros .< 1e-14)
true

FastGaussQuadrature.BESSELJ0_ROOTS — Constant

First twenty roots of Bessel funcion $J_0$ in Float64. https://mathworld.wolfram.com/BesselFunctionZeros.html

Examples

julia> zeros = besselj0.(FastGaussQuadrature.BESSELJ0_ROOTS);

julia> all(zeros .< 1e-14)
true

FastGaussQuadrature.BESSELJ1_ON_BESSELJ0_ROOTS — Constant

Values of Bessel function $J_1$ on first ten roots of Bessel function J_0.

Examples

julia> roots = approx_besselroots(0,10);

julia> (besselj1.(roots)).^2 ≈ FastGaussQuadrature.BESSELJ1_ON_BESSELJ0_ROOTS
true

FastGaussQuadrature.PIESSENS_C — Constant

Coefficients of Chebyshev series approximations for the zeros of the Bessel functions

\[j_{\nu, s} \approx \sum_{k}^{n} C_{k,s} T_k(\frac{\nu-2}{3})\]

where $j_{\nu, s}$ is a $s$-th zero of Bessel function $J_\nu$, $\{T_k\}$ are Chebyshev polynomials and $C_{k, s}$ is the coefficients.

FastGaussQuadrature.evalLaguerreRec — Method

Evaluate the orthonormal associated Laguerre polynomial with positive leading coefficient, as well as its derivative, in the point x using the recurrence relation.

FastGaussQuadrature.feval_asy1 — Method

Evaluate the interior asymptotic formula at x = cos(t). Assumption:

length(t) == n ÷ 2

FastGaussQuadrature.gausslaguerre_GW — Method

Calculate Gauss-Laguerre nodes and weights from the eigenvalue decomposition of the Jacobi matrix.

FastGaussQuadrature.gausslaguerre_asy — Method

Compute the Gauss-Laguerre rule using explicit asymptotic expansions for the nodes and weights. Optional parameters are:

reduced: compute a reduced quadrature rule, discarding all points and weights as soon as the weights underflow

T: the order of the expansion. Set T=-1 to determine the order adaptively depending on the size of the terms in the expansion

recompute: if a crude measure of the error is larger than a tolerance, the point and weight are recomputed using the (slower) recursion+newton approach, yielding more reliable accurate results.

FastGaussQuadrature.gausslaguerre_rec — Method

Compute Gauss-Laguerre rule based on the recurrence relation, using Newton iterations on an initial guess.